+++ lib/flutter_card_swiper.dart
@@ -9,3 +9,4 @@ export 'package:flutter_card_swiper/src/enums.dart';
 export 'package:flutter_card_swiper/src/properties/allowed_swipe_direction.dart';
 export 'package:flutter_card_swiper/src/typedefs.dart';
 export 'package:flutter_card_swiper/src/widget/card_swiper.dart';
+// Overlay builder typedef lives in typedefs.dart; keeping export list explicit above.